+++ AI/VCAI/Goals/VisitTile.cpp
@@ -49,8 +49,8 @@ TSubgoal VisitTile::whatToDoToAchieve()
 		}
 		else
 		{
-			return sptr(GatherArmy(fh->evaluateDanger(tile, *ret->hero) * SAFE_ATTACK_CONSTANT)
-				    .sethero(ret->hero).setisAbstract(true));
+			return sptr(GatherArmy((int)(fh->evaluateDanger(tile, *ret->hero) * SAFE_ATTACK_CONSTANT))
+						.sethero(ret->hero).setisAbstract(true));
 		}
 	}
 	return ret;
